[TODO] evalの実装

+++ eval.h
@@ -3,6 +3,6 @@
 
 #include "const.h"
 
-extern int eval(int board[BOARD_SIZE+2][BOARD_SIZE+2], int player);
+extern double eval(int board[BOARD_SIZE+2][BOARD_SIZE+2], int player);
 
 #endif
